foreign key

+++ src/main/java/br/com/etec/gabies/locadoraapietec/model/Ator.java
@@ -11,15 +11,15 @@ public class Ator {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-
-    private Long id;
+    //integer pq o professor declarou como int
+    private Integer id;
     private String nome;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
